Time: 0 ms (100.00%) | Memory: 8.7 MB (14.39%) - LeetSync

+++ 69-sqrtx/sqrtx.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
     int mySqrt(int x) {
-        return (long) sqrt(x);
+        return (int) sqrt(x);
     }
 };
